Refactor to organize using `helpers`
